Navbar added after stash saving problem
I had a Navbar inspired from the Bootstrapious code, I was fixing Maps API bugs that was not showing in the website,
in oder to keep working I decided to save in stash the code, with "git stash save -u", and apparently it broke.
After undoing the stash with the "git stash pop" command I couldn't retrieve anything that wasn't already pushed to
the Github repo, I have lost half of the pointers data in the map.js file and the navbar progresses.